unreachable to return Err(PestError::Unreachable)
Dearest Reviewer,

I found another `unreachable!()` and replaced it with `return Err(PestError::Unreachable)`. I did this in the github editor and did not test this change. I believe all the unreachables should be converted to the Err. If this was intentional please let me know.

Thank you for your time,
Dictated but not reviewed,
Becker